fixt moast bugs

+++ src/Camera.java
@@ -71,7 +71,7 @@ public class Camera {
                 Ray current = new Ray(3, this.position, pixelGrid[x][y], visible,this);
                 Vector3 color = current.castRay();
                 //System.out.println("r: " + color.x + " g: " + color.y + " b: " + color.z);
-                Color temp = new Color(color.x, color.y, color.z);
+                Color temp = new Color((int)color.x, (int)color.y, (int)color.z);
                 image.setRGB(x, y, temp.getRGB());
             }
         }
